+++ spec/features/staff/users_can_edit_an_actual_spec.rb
@@ -7,13 +7,13 @@ RSpec.feature "Users can edit an actual" do
   end
 
   context "when the user belongs to BEIS" do
-    before { authenticate!(user: user) }
     let(:user) { create(:beis_user) }
     let!(:activity) { create(:programme_activity, organisation: user.organisation) }
-    let(:report) { create(:report, :active, organisation: user.organisation, fund: activity.associated_fund) }
-    let!(:actual) { create(:actual, parent_activity: activity, report: report) }
+    let!(:actual) { create(:actual, parent_activity: activity) }
 
-    scenario "editing a actual on a programme" do
+    before { authenticate!(user: user) }
+
+    scenario "they can edit an actual on a programme but not add a comment" do
       visit organisation_activity_path(activity.organisation, activity)
 
       expect(page).to have_content(actual.value)
@@ -21,6 +21,8 @@ RSpec.feature "Users can edit an actual" do
       within("##{actual.id}") do
         click_on(t("default.link.edit"))
       end
+
+      expect(page).to_not have_field("Comment")
 
       fill_in_actual_form(
         value: "2000.51",
@@ -35,7 +37,7 @@ RSpec.feature "Users can edit an actual" do
   context "when signed in as a delivery partner" do
     let(:user) { create(:delivery_partner_user) }
     let(:activity) { create(:project_activity, organisation: user.organisation) }
-    let(:actual) { create(:actual, parent_activity: activity) }
+    let(:actual) { create(:actual, :with_comment, parent_activity: activity) }
     let(:report) { create(:report, organisation: activity.organisation, fund: activity.associated_fund) }
 
     before { authenticate!(user: user) }
@@ -55,8 +57,16 @@ RSpec.feature "Users can edit an actual" do
           expect(page).to have_content("£110.01")
           click_link("Edit")
         end
+        expect(page).to have_field("Comment", with: actual.comment.body)
 
-        fill_in "actual[value]", with: 221.12
+        fill_in "Actual amount", with: "notanumber"
+        click_on(t("default.button.submit"))
+
+        expect(page).to have_content("Value must be a valid number")
+        expect(page).to have_field("Actual amount", with: "notanumber")
+
+        fill_in "Actual amount", with: 221.12
+        fill_in "Comment", with: "Edited comment body"
 
         click_on(t("default.button.submit"))
 
@@ -64,6 +74,41 @@ RSpec.feature "Users can edit an actual" do
           expect(page).to have_content("£221.12")
         end
         expect_to_see_change_recorded_in_activitys_change_history("110.01", "221.12")
+        expect(actual.reload.comment.reload.body).to eql("Edited comment body")
+      end
+
+      context "and it doesn't have a comment" do
+        let(:actual) { create(:actual, parent_activity: activity) }
+
+        it "a comment can be added" do
+          visit organisation_activity_path(activity.organisation, activity)
+
+          within ".actuals" do
+            expect(page).to have_content("£110.01")
+            click_link("Edit")
+          end
+
+          expect(page).to have_field("Comment")
+          fill_in "Comment", with: "Fresshhh comment body"
+          click_on(t("default.button.submit"))
+
+          expect(page).to have_content("successfully")
+          expect(actual.reload.comment.body).to eql("Fresshhh comment body")
+        end
+      end
+
+      scenario "deleting the comment body deletes the comment" do
+        visit organisation_activity_path(activity.organisation, activity)
+
+        within ".actuals" do
+          click_link("Edit")
+        end
+
+        fill_in "Comment", with: ""
+        click_on(t("default.button.submit"))
+
+        expect(page).to have_content("successfully")
+        expect(actual.reload.comment).to be_nil
       end
     end