+++ CHANGELOG.md
@@ -1015,6 +1015,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Remove Auth0 user identifier from user management
 - Add sector and sector category codes to radio button text when adding a new activity
 - Fix an error where a new refund creation would silently fail when the value was non-numeric
+- Users can add and edit comments on actuals through the form
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-102...HEAD
 [release-102]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-101...release-102